wh main 입문 와인 추천 수정

+++ src/main/java/com/green/winey_final/recommend2/RecommendController2.java
@@ -42,7 +42,7 @@ public class RecommendController2 {
 
     @GetMapping("/getUserInfo")
     @Operation(summary = "유저와인정보")
-    public List<Integer> getUserInfo() {
+    public List<Long> getUserInfo() {
         return service.selUserinfo();
     }
